+++ cuda_rasterizer/rasterizer_impl.h
@@ -67,7 +67,7 @@ namespace CudaRasterizer
 	struct RegressionState
 	{
 		size_t v11v12_size;
-		float* v11v12;
+		double* v11v12;
 
 		static RegressionState fromChunk(char*& chunk, size_t P);
 	};
